updates and improvements for toolbar plugins
