+++ web/client/components/home/Home.jsx
@@ -1,0 +1,20 @@
+/**
+ * Copyright 2016, GeoSolutions Sas.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+const React = require('react');
+
+const Home = React.createClass({
+    propTypes: {},
+    getDefaultProps() {
+        return {};
+    },
+    render() {
+        return (<span/>);
+    }
+});
+
+module.exports = Home;
