+++ web/client/plugins/Home.jsx
@@ -1,0 +1,33 @@
+/**
+ * Copyright 2016, GeoSolutions Sas.
+ * All rights reserved.
+ *
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+const React = require('react');
+
+const assign = require('object-assign');
+
+const {goToPage} = require('../actions/router');
+
+const Message = require('./locale/Message');
+
+const {Glyphicon} = require('react-bootstrap');
+
+const Home = require('../components/home/Home');
+
+module.exports = {
+    HomePlugin: assign(Home, {
+        Toolbar: {
+            name: 'home',
+            position: 1,
+            tooltip: "gohome",
+            icon: <Glyphicon glyph="home"/>,
+            help: <Message msgId="helptexts.gohome"/>,
+            action: (context) => goToPage('/', context.router)
+        }
+    }),
+    reducers: {}
+};
