+++ web/client/plugins/help/HelpTextPanel.jsx
@@ -6,9 +6,26 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-const {connect} = require('react-redux');
+const React = require('react');
 
-module.exports = connect((state) => ({
+const {connect} = require('react-redux');
+const assign = require('object-assign');
+const {Glyphicon} = require('react-bootstrap');
+
+const HelpTextPanel = connect((state) => ({
     isVisible: state.controls && state.controls.help && state.controls.help.enabled,
     helpText: state.help && state.help.helpText
 }))(require('../../components/help/HelpTextPanel'));
+
+module.exports = {
+    HelpPlugin: assign(HelpTextPanel, {
+        Toolbar: {
+            name: 'help',
+            position: 1000,
+            icon: <Glyphicon glyph="question-sign"/>,
+            tooltip: "help",
+            toggle: true
+        }
+    }),
+    reducers: {}
+};
