+++ web/client/translations/data.en-US
@@ -25,6 +25,8 @@
         "mousePositionCoordinates": "Coordinates indicator",
         "menu": "Menu",
         "settings": "Settings",
+        "contactus": "Contact Us",
+        "about": "About this map",
         "help": "Help",
         "gohome": "Go home",
         "printbutton": "Print",
