+++ web/client/translations/data.it-IT
@@ -25,6 +25,8 @@
         "mousePositionCoordinates": "Indicatore di coordinate",
         "menu": "Menu",
         "settings": "Impostazioni",
+        "contactus": "Contattaci",
+        "about": "Info sulla mappa",
         "help": "Aiuto",
         "gohome": "Torna alla home",
         "printbutton": "Stampa",
